Fixed First launch app create account or import account,The chat screen blink and may cause connect mutible times
