+++ lib/screens/chat/no_wallet_account.dart
@@ -1,224 +1,232 @@
-import 'dart:convert';
-
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_easyloading/flutter_easyloading.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:nmobile/app.dart';
-import 'package:nmobile/blocs/wallet/wallets_bloc.dart';
-import 'package:nmobile/blocs/wallet/wallets_event.dart';
-import 'package:nmobile/components/box/body.dart';
-import 'package:nmobile/components/header/header.dart';
-import 'package:nmobile/components/label.dart';
-import 'package:nmobile/components/selector_text.dart';
-import 'package:nmobile/components/textbox.dart';
-import 'package:nmobile/consts/colors.dart';
-import 'package:nmobile/consts/theme.dart';
-import 'package:nmobile/helpers/local_storage.dart';
-import 'package:nmobile/helpers/secure_storage.dart';
-import 'package:nmobile/helpers/validation.dart';
-import 'package:nmobile/l10n/localization_intl.dart';
-import 'package:nmobile/plugins/nkn_wallet.dart';
-import 'package:nmobile/schemas/wallet.dart';
-import 'package:nmobile/screens/chat/authentication_helper.dart';
-import 'package:nmobile/screens/wallet/import_nkn_eth_wallet.dart';
-import 'package:nmobile/utils/extensions.dart';
-import 'package:nmobile/utils/image_utils.dart';
-
-class NoWalletAccount extends StatefulWidget {
-  static const String routeName = '/chat/no_wallet_account';
-
-  final TimerAuth timerAuth;
-
-  NoWalletAccount(this.timerAuth);
-
-  @override
-  _NoWalletAccountState createState() => _NoWalletAccountState();
-}
-
-class _NoWalletAccountState extends State<NoWalletAccount> {
-  GlobalKey _formKey = GlobalKey<FormState>();
-  bool _formValid = false;
-  TextEditingController _passwordController = TextEditingController();
-  FocusNode _nameFocusNode = FocusNode();
-  FocusNode _passwordFocusNode = FocusNode();
-  FocusNode _confirmPasswordFocusNode = FocusNode();
-  WalletsBloc _walletsBloc;
-  var _name;
-  var _password;
-
-  @override
-  void initState() {
-    super.initState();
-    _walletsBloc = BlocProvider.of<WalletsBloc>(context);
-  }
-
-  _createAccount() async {
-    if (_formValid == false){
-      return;
-    }
-    if ((_formKey.currentState as FormState).validate()) {
-      (_formKey.currentState as FormState).save();
-      EasyLoading.show();
-
-      String keystore = await NknWalletPlugin.createWallet(null, _password);
-      var json = jsonDecode(keystore);
-
-      String address = json['Address'];
-
-      await SecureStorage().set('${SecureStorage.PASSWORDS_KEY}:$address', _password);
-      await LocalStorage().saveKeyStoreInFile(address, keystore);
-      var wallet = WalletSchema(name: _name, address: address);
-
-      _walletsBloc.add(AddWallet(WalletSchema(address: address, type: WalletSchema.NKN_WALLET, name: _name), keystore));
-
-      var w = await wallet.exportWallet(_password);
-
-      EasyLoading.dismiss();
-      Navigator.of(context).pushReplacementNamed(AppScreen.routeName);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        backgroundColor: DefaultTheme.primaryColor,
-        appBar: Header(
-          titleChild: Label(NL10ns.of(context).menu_chat.toUpperCase(), type: LabelType.h2).pad(l: 20),
-          hasBack: false,
-          backgroundColor: DefaultTheme.primaryColor,
-          leading: null,
-        ),
-        body: Builder(
-            builder: (BuildContext context) => BodyBox(
-                padding: EdgeInsets.only(left: 20, right: 20),
-                color: DefaultTheme.backgroundColor1,
-                child: Center(
-                    child: SingleChildScrollView(
-                  child: Column(mainAxisAlignment: MainAxisAlignment.center, children: <Widget>[
-                    loadAssetChatPng('messages', width: 198.w, height: 144.h).pad(t: 0),
-                    Expanded(
-                      flex: 0,
-                      child: Column(
-                        children: <Widget>[
-                          Label(
-                            NL10ns.of(context).chat_no_wallet_title,
-                            type: LabelType.h2,
-                            textAlign: TextAlign.center,
-                            softWrap: true,
-                          ).pad(t: 12),
-                        ],
-                      ).pad(b: 32),
-                    ),
-                    Container(
-                        padding: 16.pad(t: 24, b: 24),
-                        decoration: BoxDecoration(
-                          color: DefaultTheme.backgroundColor2,
-                          borderRadius: BorderRadius.all(Radius.circular(32)),
-                        ),
-                        child: Form(
-                            key: _formKey,
-                            autovalidate: true,
-                            onChanged: () {
-                              setState(() {
-                                _formValid = (_formKey.currentState as FormState).validate();
-                              });
-                            },
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: <Widget>[
-                                Row(
-                                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                  children: <Widget>[
-                                    Label(
-                                      NL10ns.of(context).nickname,
-                                      type: LabelType.h3,
-                                      textAlign: TextAlign.start,
-                                    ),
-                                    TextSelector(
-                                      NL10ns.of(context).import_wallet_as_account,
-                                      DefaultTheme.bodySmallFontSize,
-                                      Colours.blue_0f,
-                                      Colours.gray_81,
-                                      fontStyle: FontStyle.italic,
-                                      decoration: TextDecoration.underline,
-                                      onTap: () {
-                                        Navigator.pushNamed(context, ImportWalletScreen.routeName, arguments: WalletType.nkn);
-                                      },
-                                    ),
-                                  ],
-                                ),
-                                Textbox(
-                                  hintText: NL10ns.of(context).input_nickname,
-                                  focusNode: _nameFocusNode,
-                                  onSaved: (v) => _name = v,
-                                  onFieldSubmitted: (_) {
-                                    FocusScope.of(context).requestFocus(_passwordFocusNode);
-                                  },
-                                  textInputAction: TextInputAction.next,
-                                  validator: Validator.of(context).walletName(),
-                                  borderColor: Colours.blue_0f,
-                                ),
-                                Label(
-                                  NL10ns.of(context).wallet_password,
-                                  type: LabelType.h3,
-                                  textAlign: TextAlign.start,
-                                ),
-                                Textbox(
-                                  focusNode: _passwordFocusNode,
-                                  controller: _passwordController,
-                                  hintText: NL10ns.of(context).input_password,
-                                  onSaved: (v) => _password = v,
-                                  onFieldSubmitted: (_) {
-                                    FocusScope.of(context).requestFocus(_confirmPasswordFocusNode);
-                                  },
-                                  textInputAction: TextInputAction.next,
-                                  validator: Validator.of(context).password(),
-                                  password: true,
-                                  padding: 0.pad(b: 8),
-                                  borderColor: Colours.blue_0f,
-                                ),
-                                Text(
-                                  NL10ns.of(context).wallet_password_mach,
-                                  style: TextStyle(color: Color(0xFF8F92A1), fontSize: DefaultTheme.bodySmallFontSize),
-                                ),
-                                Label(
-                                  NL10ns.of(context).confirm_password,
-                                  type: LabelType.h3,
-                                  textAlign: TextAlign.start,
-                                ).pad(t: 12),
-                                Textbox(
-                                  focusNode: _confirmPasswordFocusNode,
-                                  hintText: NL10ns.of(context).input_password_again,
-                                  validator: Validator.of(context).confrimPassword(_passwordController.text),
-                                  password: true,
-                                  padding: 0.pad(b: 24),
-                                  borderColor: Colours.blue_0f,
-                                ),
-                                Expanded(
-                                    flex: 0,
-                                    child: SafeArea(
-                                        child: SizedBox(
-                                      width: double.infinity,
-                                      height: 48,
-                                      child: FlatButton(
-                                        padding: const EdgeInsets.all(0),
-                                        disabledColor: Colours.light_e5,
-                                        disabledTextColor: DefaultTheme.fontColor2,
-                                        color: Colours.blue_0f,
-                                        colorBrightness: Brightness.dark,
-                                        child: Text(
-                                          NL10ns.of(context).create_account,
-                                          style: TextStyle(fontSize: DefaultTheme.h3FontSize, fontWeight: FontWeight.bold, color: null),
-                                        ),
-                                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
-                                        onPressed: _createAccount,
-                                      ),
-                                    )))
-                              ],
-                            ))),
-                  ]).pad(b: 68),
-                )))));
-  }
-}
+// import 'dart:convert';
+//
+// import 'package:flutter/material.dart';
+// import 'package:flutter/rendering.dart';
+// import 'package:flutter_bloc/flutter_bloc.dart';
+// import 'package:flutter_easyloading/flutter_easyloading.dart';
+// import 'package:flutter_screenutil/flutter_screenutil.dart';
+// import 'package:nmobile/app.dart';
+// import 'package:nmobile/blocs/chat/auth_bloc.dart';
+// import 'package:nmobile/blocs/chat/auth_event.dart';
+// import 'package:nmobile/blocs/chat/auth_state.dart';
+// import 'package:nmobile/blocs/wallet/wallets_bloc.dart';
+// import 'package:nmobile/blocs/wallet/wallets_event.dart';
+// import 'package:nmobile/components/box/body.dart';
+// import 'package:nmobile/components/header/header.dart';
+// import 'package:nmobile/components/label.dart';
+// import 'package:nmobile/components/selector_text.dart';
+// import 'package:nmobile/components/textbox.dart';
+// import 'package:nmobile/consts/colors.dart';
+// import 'package:nmobile/consts/theme.dart';
+// import 'package:nmobile/helpers/secure_storage.dart';
+// import 'package:nmobile/helpers/validation.dart';
+// import 'package:nmobile/l10n/localization_intl.dart';
+// import 'package:nmobile/plugins/nkn_wallet.dart';
+// import 'package:nmobile/schemas/wallet.dart';
+// import 'package:nmobile/screens/chat/authentication_helper.dart';
+// import 'package:nmobile/screens/wallet/import_nkn_eth_wallet.dart';
+// import 'package:nmobile/utils/extensions.dart';
+// import 'package:nmobile/utils/image_utils.dart';
+//
+// class NoWalletAccount extends StatefulWidget {
+//   static const String routeName = '/chat/no_wallet_account';
+//
+//   final TimerAuth timerAuth;
+//
+//   NoWalletAccount(this.timerAuth);
+//
+//   @override
+//   _NoWalletAccountState createState() => _NoWalletAccountState();
+// }
+//
+// class _NoWalletAccountState extends State<NoWalletAccount> {
+//   GlobalKey _formKey = GlobalKey<FormState>();
+//   bool _formValid = false;
+//   TextEditingController _passwordController = TextEditingController();
+//   FocusNode _nameFocusNode = FocusNode();
+//   FocusNode _passwordFocusNode = FocusNode();
+//   FocusNode _confirmPasswordFocusNode = FocusNode();
+//   WalletsBloc _walletsBloc;
+//   AuthBloc _authBloc;
+//   var _name;
+//   var _password;
+//
+//   @override
+//   void initState() {
+//     super.initState();
+//     _walletsBloc = BlocProvider.of<WalletsBloc>(context);
+//
+//     _authBloc = BlocProvider.of<AuthBloc>(context);
+//   }
+//
+//   _createAccount() async {
+//     if (_formValid == false){
+//       return;
+//     }
+//     if ((_formKey.currentState as FormState).validate()) {
+//       (_formKey.currentState as FormState).save();
+//       EasyLoading.show();
+//
+//       String keystore = await NknWalletPlugin.createWallet(null, _password);
+//       var json = jsonDecode(keystore);
+//
+//       String address = json['Address'];
+//
+//       await SecureStorage().set('${SecureStorage.PASSWORDS_KEY}:$address', _password);
+//       print('_createAccount__setPassword'+_password);
+//       _walletsBloc.add(AddWallet(WalletSchema(address: address, type: WalletSchema.NKN_WALLET, name: _name), keystore));
+//       _authBloc.add(AuthInProgressEvent());
+//       print('_createAccount__setKeyStore'+keystore);
+//
+//       EasyLoading.dismiss();
+//
+//       // Navigator.of(context).popUntil((route) => false)
+//     }
+//   }
+//
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//         backgroundColor: DefaultTheme.primaryColor,
+//         appBar: Header(
+//           titleChild: Label(NL10ns.of(context).menu_chat.toUpperCase(), type: LabelType.h2).pad(l: 20),
+//           hasBack: false,
+//           backgroundColor: DefaultTheme.primaryColor,
+//           leading: null,
+//         ),
+//         body: Builder(
+//             builder: (BuildContext context) => BodyBox(
+//                 padding: EdgeInsets.only(left: 20, right: 20),
+//                 color: DefaultTheme.backgroundColor1,
+//                 child: Center(
+//                     child: SingleChildScrollView(
+//                   child: Column(mainAxisAlignment: MainAxisAlignment.center, children: <Widget>[
+//                     loadAssetChatPng('messages', width: 198.w, height: 144.h).pad(t: 0),
+//                     Expanded(
+//                       flex: 0,
+//                       child: Column(
+//                         children: <Widget>[
+//                           Label(
+//                             NL10ns.of(context).chat_no_wallet_title,
+//                             type: LabelType.h2,
+//                             textAlign: TextAlign.center,
+//                             softWrap: true,
+//                           ).pad(t: 12),
+//                         ],
+//                       ).pad(b: 32),
+//                     ),
+//                     Container(
+//                         padding: 16.pad(t: 24, b: 24),
+//                         decoration: BoxDecoration(
+//                           color: DefaultTheme.backgroundColor2,
+//                           borderRadius: BorderRadius.all(Radius.circular(32)),
+//                         ),
+//                         child: Form(
+//                             key: _formKey,
+//                             autovalidate: true,
+//                             onChanged: () {
+//                               setState(() {
+//                                 _formValid = (_formKey.currentState as FormState).validate();
+//                               });
+//                             },
+//                             child: Column(
+//                               crossAxisAlignment: CrossAxisAlignment.start,
+//                               children: <Widget>[
+//                                 Row(
+//                                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
+//                                   children: <Widget>[
+//                                     Label(
+//                                       NL10ns.of(context).nickname,
+//                                       type: LabelType.h3,
+//                                       textAlign: TextAlign.start,
+//                                     ),
+//                                     TextSelector(
+//                                       NL10ns.of(context).import_wallet_as_account,
+//                                       DefaultTheme.bodySmallFontSize,
+//                                       Colours.blue_0f,
+//                                       Colours.gray_81,
+//                                       fontStyle: FontStyle.italic,
+//                                       decoration: TextDecoration.underline,
+//                                       onTap: () {
+//                                         Navigator.pushNamed(context, ImportWalletScreen.routeName, arguments: WalletType.nkn);
+//                                       },
+//                                     ),
+//                                   ],
+//                                 ),
+//                                 Textbox(
+//                                   hintText: NL10ns.of(context).input_nickname,
+//                                   focusNode: _nameFocusNode,
+//                                   onSaved: (v) => _name = v,
+//                                   onFieldSubmitted: (_) {
+//                                     FocusScope.of(context).requestFocus(_passwordFocusNode);
+//                                   },
+//                                   textInputAction: TextInputAction.next,
+//                                   validator: Validator.of(context).walletName(),
+//                                   borderColor: Colours.blue_0f,
+//                                 ),
+//                                 Label(
+//                                   NL10ns.of(context).wallet_password,
+//                                   type: LabelType.h3,
+//                                   textAlign: TextAlign.start,
+//                                 ),
+//                                 Textbox(
+//                                   focusNode: _passwordFocusNode,
+//                                   controller: _passwordController,
+//                                   hintText: NL10ns.of(context).input_password,
+//                                   onSaved: (v) => _password = v,
+//                                   onFieldSubmitted: (_) {
+//                                     FocusScope.of(context).requestFocus(_confirmPasswordFocusNode);
+//                                   },
+//                                   textInputAction: TextInputAction.next,
+//                                   validator: Validator.of(context).password(),
+//                                   password: true,
+//                                   padding: 0.pad(b: 8),
+//                                   borderColor: Colours.blue_0f,
+//                                 ),
+//                                 Text(
+//                                   NL10ns.of(context).wallet_password_mach,
+//                                   style: TextStyle(color: Color(0xFF8F92A1), fontSize: DefaultTheme.bodySmallFontSize),
+//                                 ),
+//                                 Label(
+//                                   NL10ns.of(context).confirm_password,
+//                                   type: LabelType.h3,
+//                                   textAlign: TextAlign.start,
+//                                 ).pad(t: 12),
+//                                 Textbox(
+//                                   focusNode: _confirmPasswordFocusNode,
+//                                   hintText: NL10ns.of(context).input_password_again,
+//                                   validator: Validator.of(context).confrimPassword(_passwordController.text),
+//                                   password: true,
+//                                   padding: 0.pad(b: 24),
+//                                   borderColor: Colours.blue_0f,
+//                                 ),
+//                                 Expanded(
+//                                     flex: 0,
+//                                     child: SafeArea(
+//                                         child: SizedBox(
+//                                       width: double.infinity,
+//                                       height: 48,
+//                                       child: FlatButton(
+//                                         padding: const EdgeInsets.all(0),
+//                                         disabledColor: Colours.light_e5,
+//                                         disabledTextColor: DefaultTheme.fontColor2,
+//                                         color: Colours.blue_0f,
+//                                         colorBrightness: Brightness.dark,
+//                                         child: Text(
+//                                           NL10ns.of(context).create_account,
+//                                           style: TextStyle(fontSize: DefaultTheme.h3FontSize, fontWeight: FontWeight.bold, color: null),
+//                                         ),
+//                                         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
+//                                         onPressed: _createAccount,
+//                                       ),
+//                                     )))
+//                               ],
+//                             ))),
+//                   ]).pad(b: 68),
+//                 )
+//                 )
+//             )
+//         )
+//     );
+//   }
+// }
